backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
stripe==7.11.0
PyJWT==2.8.0
psycopg2-binary==2.9.9
redis==5.0.1
aioredis==2.0.1